+++ cl_MR.py
@@ -8,7 +8,7 @@ from typing import List
 
 # Internal imports
 from cl_error import Errors, get_error_instance
-from cl_PODA import Operation, Action_Names
+from cl_PODA import Operation, Action_Names, get_PODA_instance
 from cl_DBR import Database_Record
 from func_string_manip import delete_for_sudoc
 
@@ -139,7 +139,7 @@ class Matched_Records(object):
         """Makes the request for this specific action and returns a list of IDs as a result"""
         # Action_Names based on the same connector are siilar, do not forget to udate all of them
 
-        #to avoid redundance
+        # To avoid redundance
         # Extract data
         title = delete_for_sudoc(self.local_record.utils.get_titles_as_string())
         author = delete_for_sudoc(self.local_record.utils.get_authors_as_string())
@@ -170,75 +170,88 @@ class Matched_Records(object):
                 thisTry.add_returned_ids(res.get_records_id())
                 thisTry.add_returned_records(res.get_records())
 
-        # Action isbn2ppn
-        elif action == Action_Names.ISBN2PPN:
-            i2p = id2ppn.Abes_id2ppn(webservice=id2ppn.Webservice.ISBN, useJson=True)
-            res = i2p.get_matching_ppn(self.query, check_isbn_validity=True)
-            thisTry.define_used_query(res.get_id_used())
-            if res.status != id2ppn.Id2ppn_Status.SUCCESS:
-                thisTry.error_occured(res.get_error_msg())
-            else:
-                thisTry.add_returned_ids(res.get_results(merge=True))
 
-        # Action isbn2ppn with changed ISBN
-        elif action == Action_Names.ISBN2PPN_MODIFIED_ISBN:
-            # Converting the ISBN to 10<->13
-            isbn_validity, temp, new_query = id2ppn.validate_isbn(re.sub("[^0-9X]", "", str(self.query)))
-            if isbn_validity != id2ppn.Isbn_Validity.VALID:
-                thisTry.error_occured(Errors.ISBN_MODIFICATION_FAILED)
-                return
-            if new_query[:3] != "978" and len(new_query) == 13:
-                thisTry.error_occured(Errors.ISBN_979_CAN_NOT_BE_CONVERTED)
-                return
-            new_query = pyisbn.Isbn(re.sub("[^0-9X]", "", str(new_query))).convert()
-
-            # Ensure ISBN is not empty 
-            if not new_query:
-                thisTry.error_occured(Errors.ISBN_MODIFICATION_FAILED)
+        # Action Sudoc's id2ppn webservice
+        elif action in [
+                Action_Names.ISBN2PPN,
+                Action_Names.ISBN2PPN_MODIFIED_ISBN,
+                Action_Names.ISBN2PPN_MODIFIED_ISBN_SAME_KEY,
+                Action_Names.EAN2PPN
+            ]:
+            action_instance = get_PODA_instance(action, Action_Names)
+            # Leaves if it does nto find the action 
+            if action_instance == None:
+                thisTry.error_occured(Errors.ACTION_IS_NOT_CORRECTLY_DEFINED)
                 return
 
-            # Same thing as Action ISBN2PPN
-            i2p = id2ppn.Abes_id2ppn(useJson=True)
-            res = i2p.get_matching_ppn(new_query, check_isbn_validity=True)
-            thisTry.define_used_query(res.get_id_used())
-            if res.status != id2ppn.Id2ppn_Status.SUCCESS:
-                thisTry.error_occured(res.get_error_msg())
-            else:
-                thisTry.add_returned_ids(res.get_results(merge=True))
-
-        # Action isbn2ppn with changed ISBN without recomputing th ekey
-        elif action == Action_Names.ISBN2PPN_MODIFIED_ISBN_SAME_KEY:
-            # Converting the ISBN to 10<->13 
-            new_query = None
-            if len(self.query) == 13:
-                new_query = self.query[3:]
-            elif len(self.query) == 10:
-                # Doesn't consider 979[...] as the original issue should only concern old ISBN
-                new_query = "978" + self.query
-            
-            # Ensure ISBN is not empty 
-            if not new_query:
-                thisTry.error_occured(Errors.ISBN_MODIFICATION_FAILED)
-                return
-            
-            # Same thing as Action ISBN2PPN
-            i2p = id2ppn.Abes_id2ppn(useJson=True)
-            res = i2p.get_matching_ppn(new_query, check_isbn_validity=False)
-            thisTry.define_used_query(res.get_id_used())
-            if res.status != id2ppn.Id2ppn_Status.SUCCESS:
-                thisTry.error_occured(res.get_error_msg())
-            else:
-                thisTry.add_returned_ids(res.get_results(merge=True))
-
-        # Action SRU SUdoc EAN
-        elif action == Action_Names.EAN2PPN:
-            # No EAN was found, throw an error
-            if ean == "":
+            # Action specific checks
+            new_query = ""
+            # Throw an error if it uses EAN2PPN & no EAN was found
+            if action == Action_Names.EAN2PPN and ean == "":
                 thisTry.error_occured(Errors.NO_EAN_WAS_FOUND)
                 return
-            i2p = id2ppn.Abes_id2ppn(webservice=id2ppn.Webservice.EAN, useJson=True)
-            res = i2p.get_matching_ppn(ean)
-            thisTry.define_used_query(ean)
+            # Edit the ISBN if it uses ISBN2PPN_MODIFIED_ISBN
+            elif action == Action_Names.ISBN2PPN_MODIFIED_ISBN:
+                # Converting the ISBN to 10<->13
+                isbn_validity, temp, new_query = id2ppn.validate_isbn(re.sub("[^0-9X]", "", str(self.query)))
+                if isbn_validity != id2ppn.Isbn_Validity.VALID:
+                    thisTry.error_occured(Errors.ISBN_MODIFICATION_FAILED)
+                    return
+                if new_query[:3] != "978" and len(new_query) == 13:
+                    thisTry.error_occured(Errors.ISBN_979_CAN_NOT_BE_CONVERTED)
+                    return
+                new_query = pyisbn.Isbn(re.sub("[^0-9X]", "", str(new_query))).convert()
+            # Edit the ISBN if it uses ISBN2PPN_MODIFIED_ISBN_SAME_KEY
+            elif action == Action_Names.ISBN2PPN_MODIFIED_ISBN_SAME_KEY:
+                # Converting the ISBN to 10<->13 
+                if len(self.query) == 13:
+                    new_query = self.query[3:]
+                elif len(self.query) == 10:
+                    # Doesn't consider 979[...] as the original issue should only concern old ISBN
+                    new_query = "978" + self.query
+            # Throw an error if ISBN modification end up in empty variable
+            if not new_query and action in [
+                    Action_Names.ISBN2PPN_MODIFIED_ISBN,
+                    Action_Names.ISBN2PPN_MODIFIED_ISBN_SAME_KEY
+                ]:
+                thisTry.error_occured(Errors.ISBN_MODIFICATION_FAILED)
+                return
+
+            # Defines the correct webservice (defautls to ISBN)
+            webservice = id2ppn.Webservice.ISBN
+            if action_instance.use_isbn:
+                webservice = id2ppn.Webservice.ISBN
+            elif action_instance.use_ean:
+                webservice = id2ppn.Webservice.EAN
+            
+            # Sets up the class
+            i2p = id2ppn.Abes_id2ppn(webservice=webservice, useJson=True)
+
+            # Defines the correct query
+            query = self.query
+            if action in [Action_Names.ISBN2PPN]:
+                query = self.query # I know it's useless but that way I'm don't have a doubt reading the code fast
+            elif action in [
+                    Action_Names.ISBN2PPN_MODIFIED_ISBN,
+                    Action_Names.ISBN2PPN_MODIFIED_ISBN_SAME_KEY
+                ]:
+                query = new_query
+            elif action in [Action_Names.EAN2PPN]:
+                query = ean
+
+            # Defines if it should valdiate ISBN
+            check_isbn_validity = True
+            if action in [
+                    Action_Names.ISBN2PPN_MODIFIED_ISBN_SAME_KEY,
+                    Action_Names.EAN2PPN
+                ]:
+                check_isbn_validity = False
+
+            res = i2p.get_matching_ppn(query, check_isbn_validity=check_isbn_validity)
+            thisTry.define_used_query(res.get_id_used())
+            # Tbh Idk why this one is different, but eh
+            if action == Action_Names.EAN2PPN:
+                thisTry.define_used_query(ean)
             if res.status != id2ppn.Id2ppn_Status.SUCCESS:
                 thisTry.error_occured(res.get_error_msg())
             else:
@@ -266,172 +279,107 @@ class Matched_Records(object):
                 Action_Names.SRU_SUDOC_MTI_TDO_K,
                 Action_Names.SRU_SUDOC_MTI_TDO_V
                 ]:
+            action_instance = get_PODA_instance(action, Action_Names)
+            # Leaves if it does nto find the action 
+            if action_instance == None:
+                thisTry.error_occured(Errors.ACTION_IS_NOT_CORRECTLY_DEFINED)
+                return
             sru = ssru.Sudoc_SRU()
             sru_request = []
             # TITLE
-            if action in [
-                        Action_Names.SRU_SUDOC_MTI_AUT_EDI_APU_TDO_V,
-                        Action_Names.SRU_SUDOC_MTI_AUT_APU_TDO_V,
-                        Action_Names.SRU_SUDOC_TOU_TITLE_AUTHOR_PUBLISHER_DATE_TDO_V,
-                        Action_Names.SRU_SUDOC_TOU_TITLE_AUTHOR_DATE_TDO_V,
-                        Action_Names.SRU_SUDOC_TOU_TITLE_AUTHOR_PUBLISHER_TDO_V,
-                        Action_Names.SRU_SUDOC_MTI_AUT_EDI_APU_TDO_B,
-                        Action_Names.SRU_SUDOC_MTI_AUT_APU_TDO_B,
-                        Action_Names.SRU_SUDOC_TOU_TITLE_AUTHOR_PUBLISHER_DATE_TDO_B,
-                        Action_Names.SRU_SUDOC_TOU_TITLE_AUTHOR_DATE_TDO_B,
-                        Action_Names.SRU_SUDOC_TOU_TITLE_AUTHOR_PUBLISHER_TDO_B,
-                        Action_Names.SRU_SUDOC_MTI_AUT_EDI_APU_TDO_K,
-                        Action_Names.SRU_SUDOC_MTI_AUT_APU_TDO_K,
-                        Action_Names.SRU_SUDOC_TOU_TITLE_AUTHOR_PUBLISHER_DATE_TDO_K,
-                        Action_Names.SRU_SUDOC_TOU_TITLE_AUTHOR_DATE_TDO_K,
-                        Action_Names.SRU_SUDOC_TOU_TITLE_AUTHOR_PUBLISHER_TDO_K,
-                        Action_Names.SRU_SUDOC_MTI,
-                        Action_Names.SRU_SUDOC_MTI_TDO_B,
-                        Action_Names.SRU_SUDOC_MTI_TDO_K,
-                        Action_Names.SRU_SUDOC_MTI_TDO_V
-                    ]:
+            if action_instance.use_title:
                 # Leave if empty
                 if title.strip() == "":
-                    thisTry.error_occured(Errors.REQUIRED_DATA_MISSING)
+                    thisTry.error_occured(Errors.REQUIRED_TITLE_MISSING)
                     return
-                # Else, append to query
+                # Defines the right index
+                title_index = ssru.SRU_Indexes.MTI
+                if not action_instance.specific_index:
+                    title_index = ssru.SRU_Indexes.TOU
+                # Append to query
                 sru_request.append(ssru.Part_Of_Query(
-                    ssru.SRU_Indexes.MTI,
+                    title_index,
                     ssru.SRU_Relations.EQUALS,
                     title,
                     ssru.SRU_Boolean_Operators.AND
                 ))
             # AUTHORS
-            if action in [
-                        Action_Names.SRU_SUDOC_MTI_AUT_EDI_APU_TDO_V,
-                        Action_Names.SRU_SUDOC_MTI_AUT_APU_TDO_V,
-                        Action_Names.SRU_SUDOC_TOU_TITLE_AUTHOR_PUBLISHER_DATE_TDO_V,
-                        Action_Names.SRU_SUDOC_TOU_TITLE_AUTHOR_DATE_TDO_V,
-                        Action_Names.SRU_SUDOC_TOU_TITLE_AUTHOR_PUBLISHER_TDO_V,
-                        Action_Names.SRU_SUDOC_MTI_AUT_EDI_APU_TDO_B,
-                        Action_Names.SRU_SUDOC_MTI_AUT_APU_TDO_B,
-                        Action_Names.SRU_SUDOC_TOU_TITLE_AUTHOR_PUBLISHER_DATE_TDO_B,
-                        Action_Names.SRU_SUDOC_TOU_TITLE_AUTHOR_DATE_TDO_B,
-                        Action_Names.SRU_SUDOC_TOU_TITLE_AUTHOR_PUBLISHER_TDO_B,
-                        Action_Names.SRU_SUDOC_MTI_AUT_EDI_APU_TDO_K,
-                        Action_Names.SRU_SUDOC_MTI_AUT_APU_TDO_K,
-                        Action_Names.SRU_SUDOC_TOU_TITLE_AUTHOR_PUBLISHER_DATE_TDO_K,
-                        Action_Names.SRU_SUDOC_TOU_TITLE_AUTHOR_DATE_TDO_K,
-                        Action_Names.SRU_SUDOC_TOU_TITLE_AUTHOR_PUBLISHER_TDO_K
-                    ]:
+            if action_instance.use_authors:
                 # Leave if empty
                 if author.strip() == "":
-                    thisTry.error_occured(Errors.REQUIRED_DATA_MISSING)
+                    thisTry.error_occured(Errors.REQUIRED_AUTHORS_MISSING)
                     return
-                # Else, append to query
+                # Defines the right index
+                authors_index = ssru.SRU_Indexes.AUT
+                if not action_instance.specific_index:
+                    authors_index = ssru.SRU_Indexes.TOU
+                # Append to query
                 sru_request.append(ssru.Part_Of_Query(
-                    ssru.SRU_Indexes.AUT,
+                    authors_index,
                     ssru.SRU_Relations.EQUALS,
                     author,
                     ssru.SRU_Boolean_Operators.AND
                 ))
             # PUBLISHER
-            if action in [
-                        Action_Names.SRU_SUDOC_MTI_AUT_EDI_APU_TDO_V,
-                        Action_Names.SRU_SUDOC_TOU_TITLE_AUTHOR_PUBLISHER_DATE_TDO_V,
-                        Action_Names.SRU_SUDOC_TOU_TITLE_AUTHOR_PUBLISHER_TDO_V,
-                        Action_Names.SRU_SUDOC_MTI_AUT_EDI_APU_TDO_B,
-                        Action_Names.SRU_SUDOC_TOU_TITLE_AUTHOR_PUBLISHER_DATE_TDO_B,
-                        Action_Names.SRU_SUDOC_TOU_TITLE_AUTHOR_PUBLISHER_TDO_B,
-                        Action_Names.SRU_SUDOC_MTI_AUT_EDI_APU_TDO_K,
-                        Action_Names.SRU_SUDOC_TOU_TITLE_AUTHOR_PUBLISHER_DATE_TDO_K,
-                        Action_Names.SRU_SUDOC_TOU_TITLE_AUTHOR_PUBLISHER_TDO_K
-                    ]:
+            if action_instance.use_publisher:
                 # Leave if empty
                 if publisher.strip() == "":
-                    thisTry.error_occured(Errors.REQUIRED_DATA_MISSING)
+                    thisTry.error_occured(Errors.REQUIRED_PUBLISHER_MISSING)
                     return
-                # Else, append to query
+                # Defines the right index
+                publisher_index = ssru.SRU_Indexes.EDI
+                if not action_instance.specific_index:
+                    publisher_index = ssru.SRU_Indexes.TOU
+                # Append to query
                 sru_request.append(ssru.Part_Of_Query(
-                    ssru.SRU_Indexes.EDI,
+                    publisher_index,
                     ssru.SRU_Relations.EQUALS,
                     publisher,
                     ssru.SRU_Boolean_Operators.AND
                 ))
             # DATES        
-            if action in [
-                        Action_Names.SRU_SUDOC_MTI_AUT_EDI_APU_TDO_V,
-                        Action_Names.SRU_SUDOC_MTI_AUT_APU_TDO_V,
-                        Action_Names.SRU_SUDOC_TOU_TITLE_AUTHOR_PUBLISHER_DATE_TDO_V,
-                        Action_Names.SRU_SUDOC_TOU_TITLE_AUTHOR_DATE_TDO_V,
-                        Action_Names.SRU_SUDOC_MTI_AUT_EDI_APU_TDO_B,
-                        Action_Names.SRU_SUDOC_MTI_AUT_APU_TDO_B,
-                        Action_Names.SRU_SUDOC_TOU_TITLE_AUTHOR_PUBLISHER_DATE_TDO_B,
-                        Action_Names.SRU_SUDOC_TOU_TITLE_AUTHOR_DATE_TDO_B,
-                        Action_Names.SRU_SUDOC_MTI_AUT_EDI_APU_TDO_K,
-                        Action_Names.SRU_SUDOC_MTI_AUT_APU_TDO_K,
-                        Action_Names.SRU_SUDOC_TOU_TITLE_AUTHOR_PUBLISHER_DATE_TDO_K,
-                        Action_Names.SRU_SUDOC_TOU_TITLE_AUTHOR_DATE_TDO_K
-                    ]:
+            if action_instance.use_date:
                 # Leave if empty
                 if len(dates) < 1:
-                    thisTry.error_occured(Errors.REQUIRED_DATA_MISSING)
+                    thisTry.error_occured(Errors.REQUIRED_DATE_MISSING)
                     return
-                # Else, append to query
-                sru_request.append(ssru.Part_Of_Query(
-                    ssru.SRU_Filters.APU,
-                    ssru.SRU_Relations.SUPERIOR_OR_EQUAL,
-                    oldest_date,
-                    ssru.SRU_Boolean_Operators.AND
-                ))
-                sru_request.append(ssru.Part_Of_Query(
-                    ssru.SRU_Filters.APU,
-                    ssru.SRU_Relations.INFERIOR_OR_EQUAL,
-                    newest_date,
-                    ssru.SRU_Boolean_Operators.AND
-                ))
-            # FILTERS We can use a IF - ELIF here
-            # TDO B
-            if action in [
-                        Action_Names.SRU_SUDOC_MTI_AUT_EDI_APU_TDO_B,
-                        Action_Names.SRU_SUDOC_MTI_AUT_APU_TDO_B,
-                        Action_Names.SRU_SUDOC_TOU_TITLE_AUTHOR_PUBLISHER_DATE_TDO_B,
-                        Action_Names.SRU_SUDOC_TOU_TITLE_AUTHOR_DATE_TDO_B,
-                        Action_Names.SRU_SUDOC_TOU_TITLE_AUTHOR_PUBLISHER_TDO_B,
-                        Action_Names.SRU_SUDOC_MTI_TDO_B
-                    ]:
-                sru_request.append(ssru.Part_Of_Query(
-                    ssru.SRU_Filters.TDO,
-                    ssru.SRU_Relations.EQUALS,
-                    ssru.SRU_Filter_TDO.B,
-                    ssru.SRU_Boolean_Operators.AND
-                ))
-            # TDO K
-            elif action in [
-                        Action_Names.SRU_SUDOC_MTI_AUT_EDI_APU_TDO_K,
-                        Action_Names.SRU_SUDOC_MTI_AUT_APU_TDO_K,
-                        Action_Names.SRU_SUDOC_TOU_TITLE_AUTHOR_PUBLISHER_DATE_TDO_K,
-                        Action_Names.SRU_SUDOC_TOU_TITLE_AUTHOR_DATE_TDO_K,
-                        Action_Names.SRU_SUDOC_TOU_TITLE_AUTHOR_PUBLISHER_TDO_K,
-                        Action_Names.SRU_SUDOC_MTI_TDO_K
-                    ]:
+                # Defines the right index & append to query
+                if action_instance.specific_index:
+                    sru_request.append(ssru.Part_Of_Query(
+                        ssru.SRU_Filters.APU,
+                        ssru.SRU_Relations.SUPERIOR_OR_EQUAL,
+                        oldest_date,
+                        ssru.SRU_Boolean_Operators.AND
+                    ))
+                    sru_request.append(ssru.Part_Of_Query(
+                        ssru.SRU_Filters.APU,
+                        ssru.SRU_Relations.INFERIOR_OR_EQUAL,
+                        newest_date,
+                        ssru.SRU_Boolean_Operators.AND
+                    ))
+                else:
+                    f" AND (tou={' or tou='.join([str(num) for num in dates])})",
+            # Doctype filter
+            if action_instance.use_doctype:
+                doctype_auth_val = None
+                if action_instance.doctype == "B":
+                    doctype_auth_val = ssru.SRU_Filter_TDO.B
+                elif action_instance.doctype == "K":
+                    doctype_auth_val = ssru.SRU_Filter_TDO.K
+                elif action_instance.doctype == "V":
+                    doctype_auth_val = ssru.SRU_Filter_TDO.V
+                else:
+                    # Leave if doctype is not supported
+                    thisTry.error_occured(Errors.UNSUPPORTED_DOCTYPE)
+                    return
+                # Append to query
                 sru_request.append(ssru.Part_Of_Query(
                     ssru.SRU_Filters.TDO,
                     ssru.SRU_Relations.EQUALS,
-                    ssru.SRU_Filter_TDO.K,
+                    doctype_auth_val,
                     ssru.SRU_Boolean_Operators.AND
                 ))
-            # TDO V
-            elif action in [
-                        Action_Names.SRU_SUDOC_MTI_AUT_EDI_APU_TDO_V,
-                        Action_Names.SRU_SUDOC_MTI_AUT_APU_TDO_V,
-                        Action_Names.SRU_SUDOC_TOU_TITLE_AUTHOR_PUBLISHER_DATE_TDO_V,
-                        Action_Names.SRU_SUDOC_TOU_TITLE_AUTHOR_DATE_TDO_V,
-                        Action_Names.SRU_SUDOC_TOU_TITLE_AUTHOR_PUBLISHER_TDO_V,
-                        Action_Names.SRU_SUDOC_MTI_TDO_V
-                    ]:
-                sru_request.append(ssru.Part_Of_Query(
-                    ssru.SRU_Filters.TDO,
-                    ssru.SRU_Relations.EQUALS,
-                    ssru.SRU_Filter_TDO.V,
-                    ssru.SRU_Boolean_Operators.AND
-                )) 
-            # launch search
+            # Launch search
             thisTry.define_used_query(sru.generate_query(sru_request))
             res = sru.search(
                 thisTry.query,
@@ -481,79 +429,76 @@ class Matched_Records(object):
                     Action_Names.KOHA_SRU_ANY_TITLE_PUBLISHER_DATE,
                     Action_Names.KOHA_SRU_TITLE
                 ]:
+            action_instance = get_PODA_instance(action, Action_Names)
+            # Leaves if it does nto find the action 
+            if action_instance == None:
+                thisTry.error_occured(Errors.ACTION_IS_NOT_CORRECTLY_DEFINED)
+                return
             sru = ksru.Koha_SRU(self.target_url, ksru.SRU_Version.V1_1)
             sru_request = []
             # TITLE
-            if action in [
-                        Action_Names.KOHA_SRU_TITLE_AUTHOR_PUBLISHER_DATE,
-                        Action_Names.KOHA_SRU_TITLE_AUTHOR_DATE,
-                        Action_Names.KOHA_SRU_ANY_TITLE_AUTHOR_PUBLISHER_DATE,
-                        Action_Names.KOHA_SRU_ANY_TITLE_AUTHOR_DATE,
-                        Action_Names.KOHA_SRU_ANY_TITLE_PUBLISHER_DATE,
-                        Action_Names.KOHA_SRU_TITLE
-                    ]:
+            if action_instance.use_title:
                 # Leave if empty
                 if title.strip() == "":
-                    thisTry.error_occured(Errors.REQUIRED_DATA_MISSING)
+                    thisTry.error_occured(Errors.REQUIRED_TITLE_MISSING)
                     return
-                # Else, append to query
+                # Defines the right index
+                title_index = ksru.SRU_Indexes.TITLE
+                if not action_instance.specific_index:
+                    title_index = ksru.SRU_Indexes.ANY
+                # Append to query
                 sru_request.append(ksru.Part_Of_Query(
-                    ksru.SRU_Indexes.TITLE,
+                    title_index,
                     ksru.SRU_Relations.EQUALS,
                     title,
                     ksru.SRU_Boolean_Operators.AND
                 ))
             # AUTHORS
-            if action in [
-                        Action_Names.KOHA_SRU_TITLE_AUTHOR_PUBLISHER_DATE,
-                        Action_Names.KOHA_SRU_TITLE_AUTHOR_DATE,
-                        Action_Names.KOHA_SRU_ANY_TITLE_AUTHOR_PUBLISHER_DATE,
-                        Action_Names.KOHA_SRU_ANY_TITLE_AUTHOR_DATE
-                    ]:
+            if action_instance.use_authors:
                 # Leave if empty
                 if author.strip() == "":
-                    thisTry.error_occured(Errors.REQUIRED_DATA_MISSING)
+                    thisTry.error_occured(Errors.REQUIRED_AUTHORS_MISSING)
                     return
-                # Else, append to query
-                sru_request.append(ksru.Part_Of_Query(
-                    ksru.SRU_Indexes.AUTHOR,
+                # Defines the right index
+                authors_index = ksru.SRU_Indexes.AUTHOR
+                if not action_instance.specific_index:
+                    authors_index = ksru.SRU_Indexes.ANY
+                # Append to query
+                sru_request.append(ssru.Part_Of_Query(
+                    authors_index,
                     ksru.SRU_Relations.EQUALS,
                     author,
                     ksru.SRU_Boolean_Operators.AND
                 ))
             # PUBLISHER
-            if action in [
-                        Action_Names.KOHA_SRU_TITLE_AUTHOR_PUBLISHER_DATE,
-                        Action_Names.KOHA_SRU_ANY_TITLE_AUTHOR_PUBLISHER_DATE,
-                        Action_Names.KOHA_SRU_ANY_TITLE_PUBLISHER_DATE
-                    ]:
+            if action_instance.use_publisher:
                 # Leave if empty
                 if publisher.strip() == "":
-                    thisTry.error_occured(Errors.REQUIRED_DATA_MISSING)
+                    thisTry.error_occured(Errors.REQUIRED_PUBLISHER_MISSING)
                     return
-                # Else, append to query
+                # Defines the right index
+                publisher_index = ksru.SRU_Indexes.PUBLISHER
+                if not action_instance.specific_index:
+                    publisher_index = ksru.SRU_Indexes.ANY
+                # Append to query
                 sru_request.append(ksru.Part_Of_Query(
-                    ksru.SRU_Indexes.PUBLISHER,
+                    publisher_index,
                     ksru.SRU_Relations.EQUALS,
                     publisher,
                     ksru.SRU_Boolean_Operators.AND
                 ))
             # DATES        
-            if action in [
-                        Action_Names.KOHA_SRU_TITLE_AUTHOR_PUBLISHER_DATE,
-                        Action_Names.KOHA_SRU_TITLE_AUTHOR_DATE,
-                        Action_Names.KOHA_SRU_ANY_TITLE_AUTHOR_PUBLISHER_DATE,
-                        Action_Names.KOHA_SRU_ANY_TITLE_AUTHOR_DATE,
-                        Action_Names.KOHA_SRU_ANY_TITLE_PUBLISHER_DATE
-                    ]:
+            if action_instance.use_date:
                 # Leave if empty
                 if len(dates) < 1:
-                    thisTry.error_occured(Errors.REQUIRED_DATA_MISSING)
+                    thisTry.error_occured(Errors.REQUIRED_DATE_MISSING)
                     return
-                # Else, append to query
-                sru_request.append(f" and ({ksru.SRU_Indexes.DATE.value}={f' or {ksru.SRU_Indexes.DATE.value}='.join([str(num) for num in dates])})")
-            
-            # launch search
+                # Defines the right index
+                date_index = ksru.SRU_Indexes.DATE
+                if not action_instance.specific_index:
+                    date_index = ksru.SRU_Indexes.ANY
+                # Append to query
+                sru_request.append(f" and ({date_index.value}={f' or {date_index.value}='.join([str(num) for num in dates])})")
             thisTry.define_used_query(sru.generate_query(sru_request))
             res = sru.search(
                 thisTry.query,